+++ docs/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Changed
+- Make sure the unacked PS update stream is always linked to PS module (OSIDB-637)
+
 ## [2.3.3] - 2022-12-13
 ### Changed
 - Link unacked PS update stream to PS module on product definitions sync (OSIDB-629)